Remove check for mutable vars on objects
We should mimic the behaviour of Corda here, which does the following:
`They are recorded into the stream with no properties, and deserialize
back to the singleton instance.`
https://docs.corda.net/docs/corda-os/4.8/serialization.html#kotlin-objects

This is not unreasonable. In other well-known serialization libs like Jackson,
The approach for this has been a long debate and the result is that
how mutable state of objects is handled is configurable. But in no
case do they fail serialization if mutable state is encountered.

We could possibly choose to emit a log warning here,
but it seems we have nog logging mechanism configured.